# Reimplement Source::indexOf(ByteString) without Source::peek calls