As in https://github.com/Kotlin/kotlinx-io/pull/240, reimplemented algorithm without the use of the `PeekSource`. Also added a specialized implementation for `Buffer`.