@fzhinkin 2023-11-28
Here are benchmarking results (I executed benchmarks in throughput mode, so higher the score, better the performance):

baseline (devel):
```
Benchmark                    (params)   Mode  Cnt       Score      Error  Units
IndexOfByteString.benchmark    1024:2  thrpt    5  171907.309 ± 6796.564  ops/s
IndexOfByteString.benchmark    8192:2  thrpt    5   19421.300 ±  401.089  ops/s
IndexOfByteString.benchmark   10000:2  thrpt    5   16095.808 ±  163.348  ops/s
IndexOfByteString.benchmark   10000:8  thrpt    5    8474.794 ±   24.459  ops/s
```

new version (this branch):
```
Benchmark                    (params)   Mode  Cnt       Score       Error  Units
IndexOfByteString.benchmark    1024:2  thrpt    5  534757.836 ± 11554.425  ops/s
IndexOfByteString.benchmark    8192:2  thrpt    5   67568.970 ±   244.761  ops/s
IndexOfByteString.benchmark   10000:2  thrpt    5   51233.638 ±   365.530  ops/s
IndexOfByteString.benchmark   10000:8  thrpt    5   30259.195 ±   421.080  ops/s
```